Return exact matching keys from new endpoint
