+++ typescript/packages/toolshed/routes/ai/spell/spell.handlers.ts
@@ -9,6 +9,8 @@ import { performSearch } from "./behavior/search.ts";
 import { checkSchemaMatch } from "@/lib/schema-match.ts";
 import { Logger } from "@/lib/prefixed-logger.ts";
 import { processSchema } from "@/routes/ai/spell/fulfill.ts";
+import { candidates } from "@/routes/ai/spell/caster.ts";
+import { CasterSchemaRoute } from "@/routes/ai/spell/spell.routes.ts";
 
 // Process Schema schemas
 export const ProcessSchemaRequestSchema = z.object({
@@ -88,6 +90,13 @@ export const SearchSchemaResponseSchema = z.object({
 export type SearchSchemaRequest = z.infer<typeof SearchSchemaRequestSchema>;
 export type SearchSchemaResponse = z.infer<typeof SearchSchemaResponseSchema>;
 
+export const CasterResponseSchema = z.object({
+  spells: z.array(z.string()),
+  keys: z.array(z.string()),
+});
+
+export type CasterResponse = z.infer<typeof CasterResponseSchema>;
+
 export const fulfill: AppRouteHandler<ProcessSchemaRoute> = async (c) => {
   const logger: Logger = c.get("logger");
   const body = (await c.req.json()) as ProcessSchemaRequest;
@@ -127,6 +136,31 @@ export const search: AppRouteHandler<SearchSchemaRoute> = async (c) => {
     logger.error({ error }, "Error processing search");
     return c.json(
       { error: "Failed to process search" },
+      HttpStatusCodes.INTERNAL_SERVER_ERROR,
+    );
+  }
+};
+
+export const caster: AppRouteHandler<CasterSchemaRoute> = async (c) => {
+  const logger: Logger = c.get("logger");
+  const body = (await c.req.json()) as ProcessSchemaRequest;
+  const startTime = performance.now();
+
+  try {
+    const blobContents = await getAllBlobs(true) as Record<
+      string,
+      Record<string, unknown>
+    >;
+    const response = await candidates(body.schema, blobContents);
+
+    return c.json(
+      { keys: response.map((r) => r.key), spells: [] },
+      HttpStatusCodes.OK,
+    );
+  } catch (error) {
+    logger.error({ error }, "Error processing schema");
+    return c.json(
+      { error: "Failed to process schema" },
       HttpStatusCodes.INTERNAL_SERVER_ERROR,
     );
   }
