+++ typescript/packages/toolshed/routes/ai/spell/spell.routes.ts
@@ -2,6 +2,7 @@ import { createRoute } from "@hono/zod-openapi";
 import * as HttpStatusCodes from "stoker/http-status-codes";
 import { jsonContent } from "stoker/openapi/helpers";
 import {
+  CasterResponseSchema,
   ProcessSchemaRequestSchema,
   ProcessSchemaResponseSchema,
   SearchSchemaRequestSchema,
@@ -68,3 +69,30 @@ export const search = createRoute({
 });
 
 export type SearchSchemaRoute = typeof search;
+
+export const caster = createRoute({
+  path: "/ai/spell/caster",
+  method: "post",
+  tags,
+  request: {
+    body: {
+      content: {
+        "application/json": {
+          schema: ProcessSchemaRequestSchema,
+        },
+      },
+    },
+  },
+  responses: {
+    [HttpStatusCodes.OK]: jsonContent(
+      CasterResponseSchema,
+      "The caster results",
+    ),
+    [HttpStatusCodes.INTERNAL_SERVER_ERROR]: jsonContent(
+      ErrorResponseSchema,
+      "An error occurred",
+    ),
+  },
+});
+
+export type CasterSchemaRoute = typeof caster;
